pyproject.toml
[project]
name = "glom"
version = "0.1.0"
description = "Index and search agent context from ~/.claude and ~/.codex"
requires-python = ">=3.11"
dependencies = [
    "rich-click>=1.8",
    "pyyaml>=6.0",
]

[project.scripts]
glom = "glom.cli:main"

[dependency-groups]
dev = ["pytest>=7.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
